ci-scripts: build: docker-build: revamp container labels
To make labels more useful for users and devs alike, use a common 'torizon.'
top level domain and add the Debian Snapshot as a label, so it's easy to track
which Toradex Debian Snapshot is included within each image.

One can use `docker inspect` to get this information from an image.

Signed-off-by: Leonardo Held <leonardo.held@toradex.com>