build: refactor rules for when to trigger build jobs
Build jobs were once a upon a time conditioned to the changes done in certain
Dockerfiles or directory. We had the expectation that building only the
necessary jobs that would lead to merge requests being merged faster.

After a refactor from the gigantic yaml build files to smaller ones in the
current 'per-platform' basis, these rules were not applicable anymore, so
remove them.

Signed-off-by: Leonardo Held <leonardo.held@toradex.com>